requirements.txt
# flask
Flask==1.1.2
Flask-Compress==1.4.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-bcrypt==0.7.1
flask-sqlalchemy==2.5.1
flask-mail==0.9.1
flask-api-cache==0.1.1

# db
PyMySQL==0.9.2
SQLAlchemy==1.3.23

# redis
redis==3.0.1
hiredis==1.0.0

# others
gunicorn==20.1.0
schema==0.7.2
gevent==21.1.2
Werkzeug==0.15.5
itsdangerous==1.1.0
PyJWT==1.6.4
loguru==0.6.0